+++ spec/determinator/tracking/tracker_spec.rb
@@ -22,10 +22,12 @@ describe Determinator::Tracking::Tracker do
 
   describe '#finish!' do
     let(:feature) { FactoryGirl.build(:feature, name: 'test_feature') }
-    let(:perform) { subject.finish!(error: true, foo: :bar) }
+    let(:perform) { subject.finish!(endpoint: 'test', error: true, foo: :bar) }
+    let(:time)    { Time.now }
 
     before do
       allow(Process).to receive(:clock_gettime).and_return(1.0, 3.0)
+      allow(Time).to receive(:now).and_return(time)
       subject.track(123, 'abc', feature, 'A')
     end
 
@@ -33,8 +35,9 @@ describe Determinator::Tracking::Tracker do
       expect(perform).to be_a(Determinator::Tracking::Request)
     end
 
-    specify { expect(perform.start).to eq(1.0)}
+    specify { expect(perform.start).to eq(time)}
     specify { expect(perform.type).to eq(:test) }
+    specify { expect(perform.endpoint).to eq('test') }
     specify { expect(perform.time).to eq(2.0) }
     specify { expect(perform.error).to eq(true) }
     specify { expect(perform.attributes).to eq({foo: :bar}) }